+++ src/serac/numerics/functional/domain_integral.hpp
@@ -63,7 +63,7 @@ public:
     integral_to_functional_ = active_arguments;
     functional_to_integral_ = std::vector<int>(num_trial_spaces, -1);
     for (size_t i = 0; i < active_arguments.size(); i++) {
-      functional_to_integral_[active_arguments[i]] = static_cast<int>(i);
+      functional_to_integral_[static_cast<size_t>(active_arguments[i])] = static_cast<int>(i);
     }
 
     SERAC_MARK_BEGIN("Domain Integral Set Up");
@@ -169,7 +169,7 @@ public:
       selected[i] = &input_E[size_t(integral_to_functional_[i])];
     }
 
-    int index = (functional_index == -1) ? -1 : functional_to_integral_[functional_index];
+    int index = (functional_index == -1) ? -1 : functional_to_integral_[static_cast<size_t>(functional_index)];
     if (index == -1) {
       SERAC_MARK_BEGIN("Domain Integral Evaluation");
       evaluation_(selected, output_E, update_state);
